rendu

+++ srcs/libft/ft_strlen.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alesteph <alesteph@student.42.fr>          +#+  +:+       +#+        */
+/*   By: trmonich <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2018/11/06 15:12:09 by alesteph          #+#    #+#             */
-/*   Updated: 2018/11/06 16:28:12 by alesteph         ###   ########.fr       */
+/*   Created: 2018/11/07 09:28:49 by trmonich          #+#    #+#             */
+/*   Updated: 2018/11/08 10:19:45 by trmonich         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned long	i;
+	int		i;
 
 	i = 0;
 	while (s[i])
